requirements.txt
phe==1.5.0
pycryptodome==3.19.1
ttkbootstrap==1.10.1